added text to image core

+++ Assets/Gaxos Content Generator/ContentGeneration/Models/Stability/StabilityParameters.cs
@@ -14,7 +14,7 @@ namespace ContentGeneration.Models.Stability
         public Sampler? Sampler;
 
         [JsonProperty("samples")] public uint Samples = 1;
-        [JsonProperty("seed")] public uint Seed;
+        [JsonProperty("seed")] public ulong Seed;
         [JsonProperty("steps")] public uint Steps = 30;
 
         [JsonProperty("style_preset")]
